backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Settings
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# EPUB Processing
ebooklib==0.18
Pillow==11.1.0
beautifulsoup4==4.12.3
lxml==5.3.0

# AI & ML
sentence-transformers==3.3.1
scikit-surprise==1.1.4
xgboost==2.1.3
scikit-learn==1.6.1
networkx==3.4.2
joblib==1.4.2
pandas==2.2.3

# TTS
edge-tts==7.2.8

# Vocabulary
nltk==3.9.1

# Vector DB
pgvector==0.3.6
chromadb==0.5.23

# HTTP Client
httpx==0.28.1

# Utils
python-dotenv==1.0.1
aiofiles==24.1.0
